+++ io-core/src/com/fleeksoft/io/CharArrayReader.kt
