+++ io-core/src/com/fleeksoft/io/FilterInputStream.kt
