+++ io-core/src/com/fleeksoft/io/FilterOutputStream.kt
