+++ io-core/src/com/fleeksoft/io/FilterReader.kt
