+++ io-core/src@jvmAndAndroid/com/fleeksoft/io/CharArrayReader.jvmAndAndroid.kt
@@ -1,0 +1,5 @@
+package com.fleeksoft.io
+
+import java.io.CharArrayReader
+
+actual typealias CharArrayReader = CharArrayReader
