+++ io-core/src@jvmAndAndroid/com/fleeksoft/io/FilterInputStream.jvmAndAndroid.kt
@@ -1,0 +1,5 @@
+package com.fleeksoft.io
+
+import java.io.FilterInputStream
+
+actual typealias FilterInputStream = FilterInputStream
