+++ io-core/src@nonJvm/com/fleeksoft/io/BufferedReader.nonJvm.kt
