--- io/src@jvmAndAndroid/com/fleeksoft/io/CharArrayReader.jvmAndAndroid.kt
@@ -1,3 +1,0 @@
-package com.fleeksoft.io
-
-actual typealias CharArrayReader = java.io.CharArrayReader
