--- io/src@jvmAndAndroid/com/fleeksoft/io/FilterInputStream.jvmAndAndroid.kt
@@ -1,3 +1,0 @@
-package com.fleeksoft.io
-
-actual typealias FilterInputStream = java.io.FilterInputStream
